fix: Add performance timings clear before and during the game
It fills up as the games does on but has a maximum buffer size. The buffer could be increased but I prefer this way as it is more robust.

This fix should only affect the training mod.
As there are 3 requests per question, and the training mod questions for all users, the buffer is exceeded at least once.